Changed up the change events.
songChange has been replaced with a new method "onChange('type', callback)".
It now takes a type of callback, such as 'song' or 'playlist'.
'song' events trigger when the current song changes to a different one.
'playlist' events trigger when the playlist mutates, allowing you to easily update your view when
the playlist changes.