New caching feature
Why send a request for the same file every time?
netRequest, now named "fetch", now stores the file data from XHRs in a fetched object.
Every time you make a request for song data using a URL, fetch looks to see if it already has that data and returns it if it does, otherwise it just sends a request for it.